+++ lib/brando/blueprint/utils.ex
@@ -16,15 +16,11 @@ defmodule Brando.Blueprint.Utils do
     :sort_param,
     :with
   ]
-  @strip_ecto_opts @changeset_opts ++
-                     [
-                       :cast,
-                       :constraint_name,
-                       :constraints,
-                       :module,
-                       :rename_from,
-                       :unique
-                     ]
+  @migration_opts [:null, :precision, :scale]
+  @attribute_strip_opts @migration_opts ++ [:constraints, :rename_from, :required, :unique]
+  @relation_strip_opts @changeset_opts ++
+                         @migration_opts ++
+                         [:cast, :constraint_name, :constraints, :module, :rename_from, :unique]
   @status_type Module.concat(["Brando", "Type", "Status"])
   @file_type Module.concat(["Brando", "Type", "File"])
   @image_type Module.concat(["Brando", "Type", "Image"])
@@ -35,6 +31,7 @@ defmodule Brando.Blueprint.Utils do
   Maps a Blueprint attribute type to its Ecto schema type.
   """
   @spec to_ecto_type(term()) :: term()
+  def to_ecto_type({:array, type}), do: {:array, to_ecto_type(type)}
   def to_ecto_type(:text), do: :string
   def to_ecto_type(:status), do: @status_type
   def to_ecto_type(:file), do: @file_type
@@ -59,26 +56,30 @@ defmodule Brando.Blueprint.Utils do
   Removes Blueprint-only metadata and returns options accepted by Ecto schema macros.
   """
   @spec to_ecto_opts(term(), map()) :: keyword()
-  def to_ecto_opts(:language, opts), do: Map.to_list(opts)
+  def to_ecto_opts(:language, opts) do
+    opts
+    |> Map.drop([:languages | @attribute_strip_opts])
+    |> Map.to_list()
+  end
 
   def to_ecto_opts(:belongs_to, opts),
-    do: opts |> Map.drop(@strip_ecto_opts ++ [:on_delete]) |> Map.to_list()
+    do: opts |> Map.drop(@relation_strip_opts ++ [:on_delete]) |> Map.to_list()
 
-  def to_ecto_opts(:has_one, opts), do: opts |> Map.drop(@strip_ecto_opts) |> Map.to_list()
-  def to_ecto_opts(:many_to_many, opts), do: opts |> Map.drop(@strip_ecto_opts) |> Map.to_list()
-  def to_ecto_opts(:has_many, opts), do: opts |> Map.drop(@strip_ecto_opts) |> Map.to_list()
+  def to_ecto_opts(:has_one, opts), do: opts |> Map.drop(@relation_strip_opts) |> Map.to_list()
+  def to_ecto_opts(:many_to_many, opts), do: opts |> Map.drop(@relation_strip_opts) |> Map.to_list()
+  def to_ecto_opts(:has_many, opts), do: opts |> Map.drop(@relation_strip_opts) |> Map.to_list()
 
   def to_ecto_opts(:embeds_one, opts) do
     opts
     |> Map.put_new(:on_replace, :update)
-    |> Map.drop(@strip_ecto_opts)
+    |> Map.drop(@relation_strip_opts)
     |> Map.to_list()
   end
 
   def to_ecto_opts(:embeds_many, opts) do
     opts
     |> Map.put_new(:on_replace, :delete)
-    |> Map.drop(@strip_ecto_opts)
+    |> Map.drop(@relation_strip_opts)
     |> Map.to_list()
   end
 
@@ -86,7 +87,7 @@ defmodule Brando.Blueprint.Utils do
     opts
     |> Map.put(:array?, false)
     |> Map.put(:default, nil)
-    |> Map.drop(@strip_ecto_opts)
+    |> Map.drop(@attribute_strip_opts)
     |> Map.to_list()
   end
 
@@ -94,11 +95,11 @@ defmodule Brando.Blueprint.Utils do
     opts
     |> Map.put(:array?, true)
     |> Map.put(:default, [])
-    |> Map.drop(@strip_ecto_opts)
+    |> Map.drop(@attribute_strip_opts)
     |> Map.to_list()
   end
 
-  def to_ecto_opts(_type, opts), do: opts |> Map.drop(@strip_ecto_opts) |> Map.to_list()
+  def to_ecto_opts(_type, opts), do: opts |> Map.drop(@attribute_strip_opts) |> Map.to_list()
 
   @doc """
   Removes Blueprint-only metadata and returns options accepted by Ecto cast helpers.